meta: move _errors to errors with related error classes
