+++ snapcraft/internal/errors.py
@@ -162,85 +162,6 @@ class PrimeFileConflictError(SnapcraftError):
     )
 
 
-class InvalidAppCommandError(SnapcraftError):
-
-    fmt = (
-        "Failed to generate snap metadata: "
-        "The specified command {command!r} defined in the app {app_name!r} does "
-        "not exist or is not executable.\n"
-        "Ensure that {command!r} is installed with the correct path."
-    )
-
-    def __init__(self, command, app_name):
-        super().__init__(command=command, app_name=app_name)
-
-
-class InvalidAppCommandNotFound(SnapcraftError):
-
-    fmt = (
-        "Failed to generate snap metadata: "
-        "The specified command {command!r} defined in the app {app_name!r} does "
-        "not exist.\n"
-        "Ensure that {command!r} is installed with the correct path."
-    )
-
-    def __init__(self, command, app_name):
-        super().__init__(command=command, app_name=app_name)
-
-
-class InvalidAppCommandNotExecutable(SnapcraftError):
-
-    fmt = (
-        "Failed to generate snap metadata: "
-        "The specified command {command!r} defined in the app {app_name!r} "
-        "is not executable."
-    )
-
-    def __init__(self, command, app_name):
-        super().__init__(command=command, app_name=app_name)
-
-
-class InvalidAppCommandFormatError(SnapcraftError):
-
-    fmt = (
-        "Failed to generate snap metadata: "
-        "The specified command {command!r} defined in the app {app_name!r} does "
-        "not match the pattern expected by snapd.\n"
-        "The command must consist only of alphanumeric characters, spaces, and the "
-        "following special characters: / . _ # : $ -"
-    )
-
-    def __init__(self, command, app_name):
-        super().__init__(command=command, app_name=app_name)
-
-
-class InvalidCommandChainError(SnapcraftError):
-
-    fmt = (
-        "Failed to generate snap metadata: "
-        "The command-chain item {item!r} defined in the app {app_name!r} does "
-        "not exist or is not executable.\n"
-        "Ensure that {item!r} is relative to the prime directory."
-    )
-
-    def __init__(self, item, app_name):
-        super().__init__(item=item, app_name=app_name)
-
-
-class InvalidDesktopFileError(SnapcraftError):
-
-    fmt = (
-        "Failed to generate desktop file: "
-        "Invalid desktop file {filename!r}: {message}."
-        # FIXME include how to fix each of the possible desktop file errors.
-        # https://bugs.launchpad.net/snapcraft/+bug/1727435
-        # --elopio - 2017-10-25
-    )
-
-    def __init__(self, filename, message):
-        super().__init__(filename=filename, message=message)
-
-
 class PluginError(SnapcraftError):
 
     fmt = (
